requirements.txt
configparser==3.5.0
progressbar2==3.12.0
python-utils==2.0.1
sh==1.12.10
six==1.10.0
